FROMLIST: staging: apfs: init TODO and README.rst
There are various cleanup tasks required before this driver can be moved
to fs/ and out of staging/. TODO attempts to start a list of what those
tasks may be.

The README delineates a history and provides credit to the module's
original authors. I am not the original author, I have just ported it
for upstream submission.

Signed-off-by: Ethan Carter Edwards <ethan@ethancedwards.com>

Link: https://lore.kernel.org/all/20250319-apfs-v2-0-475de2e25782@ethancedwards.com/
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>