iommu/dart: Add iommu_ops for locked DARTs
A locked DART has partially read-only MMIO registers. Most importantly
the TTBR registers are read-only. Apple's bootloader sets the DART up
for its intended use before locking it. The single used streams has
a L1 translation table allocated in carved out memory and its TTBRs
point to this table.
In addition translation and bypass can not be disabled or
enabled so a locked DART must not offer default identity or blocked
domains.

The only observed locked DART is for the display coprocessor. It
requires careful handling as translation errors result in unrecoverable
crashes of the display coprocessor.

Signed-off-by: Janne Grunau <j@jannau.net>